Cursor Inactivity/requirements.txt
requests>=2.25.0
pandas>=1.3.0
PyYAML>=5.4.0